requirements.txt
fastapi
python-can